Let remaining time move on to the next bus time
Previously, if the tick is down to zero, it would stay at zero. But now, it would automatically pull up the next bus time and tick again. Even though it seems great, there is two issue with this: one is that if the last bus for some route ticks zero, it would change the remaining time to NaN. I still can't find why. Second, the method I used to pull up the bus schedule only works for the day that I accessed this page.(11/19 23:40 -> All the bus schedule will show up 0 because there is no bus available in 11/19.)

These two issues need to be fixed.